# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Proector API",
        "description": "Backend API documentation for Proector mobile and admin platforms.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API base path"
        },
        {
            "url": "http://localhost"
        }
    ],
    "paths": {
        "/api/admin/login": {
            "post": {
                "tags": [
                    "Admin Auth"
                ],
                "summary": "Login admin",
                "description": "Authenticates admin with username/password and returns scoped access token.",
                "operationId": "02064f557af8baf1c24fd21c09b689f3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "password"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful"
                    }
                }
            }
        },
        "/api/admin/logout": {
            "post": {
                "tags": [
                    "Admin Auth"
                ],
                "summary": "Logout admin",
                "description": "Revokes current admin access token.",
                "operationId": "18687458536b64e1e1c3725706529bb2",
                "responses": {
                    "200": {
                        "description": "Logged out"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/bookings": {
            "get": {
                "tags": [
                    "Admin Bookings"
                ],
                "summary": "Get all bookings",
                "description": "Returns paginated bookings list for admin operations with optional status filter.",
                "operationId": "2259635a015b5c8d898d2a8776cdcca0",
                "responses": {
                    "200": {
                        "description": "Bookings list"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/bookings/{id}": {
            "get": {
                "tags": [
                    "Admin Bookings"
                ],
                "summary": "Get booking details",
                "description": "Returns full booking details for admin review and management.",
                "operationId": "0e796f08e36f923c8ba17fc6f8af609b",
                "responses": {
                    "200": {
                        "description": "Booking details"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Admin Bookings"
                ],
                "summary": "Update booking",
                "description": "Updates editable booking attributes such as start time, duration, team, and notes.",
                "operationId": "9a2a41ef271ab85e144ed93b08d7d881",
                "responses": {
                    "200": {
                        "description": "Booking updated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/bookings/{id}/assign-team": {
            "post": {
                "tags": [
                    "Admin Bookings"
                ],
                "summary": "Assign security team to booking",
                "description": "Assigns an available team to pending booking and transitions status to confirmed.",
                "operationId": "b4cfa43fcf8aa7431302910eabb293e5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "security_team_id"
                                ],
                                "properties": {
                                    "security_team_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Team assigned"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/bookings/{id}/complete": {
            "post": {
                "tags": [
                    "Admin Bookings"
                ],
                "summary": "Complete booking by admin",
                "description": "Completes booking from admin side and releases assigned team availability.",
                "operationId": "a4c8dea515a0b260ae799dd72330dc67",
                "responses": {
                    "200": {
                        "description": "Booking completed"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/clients": {
            "get": {
                "tags": [
                    "Admin Clients"
                ],
                "summary": "Get all clients",
                "description": "Returns client list for admin with verification and activity details.",
                "operationId": "6f606ceba2fad5c8ee58c894625ffdd9",
                "responses": {
                    "200": {
                        "description": "Clients list"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/clients/{id}": {
            "get": {
                "tags": [
                    "Admin Clients"
                ],
                "summary": "Get client details",
                "description": "Returns full client profile including verification docs and booking/payment history.",
                "operationId": "593c12d3d04096cc17a8e244177eef9c",
                "responses": {
                    "200": {
                        "description": "Client details"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/clients/{id}/verify": {
            "post": {
                "tags": [
                    "Admin Clients"
                ],
                "summary": "Verify or reject client verification",
                "description": "Approves or rejects client verification request with optional rejection reason.",
                "operationId": "2eaf5ef5e88099b4a5b11119126f4a01",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "verification_status"
                                ],
                                "properties": {
                                    "verification_status": {
                                        "type": "string",
                                        "enum": [
                                            "verified",
                                            "rejected"
                                        ]
                                    },
                                    "rejection_reason": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Verification status updated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Use Bearer {token} from login or OTP verification endpoints.",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}